moved tests to a separate test file.
